+++ src/main/java/uk/gov/companieshouse/itemgroupstatusupdater/config/Config.java
@@ -3,8 +3,8 @@ package uk.gov.companieshouse.itemgroupstatusupdater.config;
 import static uk.gov.companieshouse.itemgroupstatusupdater.ItemGroupStatusUpdaterApplication.NAMESPACE;
 
 import consumer.deserialization.AvroDeserializer;
-import uk.gov.companieshouse.itemgroupstatusupdater.InvalidMessageRouter;
-import uk.gov.companieshouse.itemgroupstatusupdater.MessageFlags;
+import uk.gov.companieshouse.itemgroupstatusupdater.kafka.InvalidMessageRouter;
+import uk.gov.companieshouse.itemgroupstatusupdater.kafka.MessageFlags;
 import uk.gov.companieshouse.logging.util.DataMap;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.producer.ProducerConfig;
