+++ src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/kafka/AbstractKafkaIntegrationTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.itemgroupstatusupdater;
+package uk.gov.companieshouse.itemgroupstatusupdater.kafka;
 
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -7,6 +7,7 @@ import org.testcontainers.containers.KafkaContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
+import uk.gov.companieshouse.itemgroupstatusupdater.TestConfig;
 
 @Testcontainers
 @Import(TestConfig.class)
