+++ src/test/java/uk/gov/companieshouse/itemgroupstatusupdater/kafka/ConsumerInvalidTopicTest.java
@@ -1,16 +1,16 @@
-package uk.gov.companieshouse.itemgroupstatusupdater;
+package uk.gov.companieshouse.itemgroupstatusupdater.kafka;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.verify;
 import static uk.gov.companieshouse.itemgroupstatusupdater.Constants.ITEM_GROUP_PROCESSED;
 import static uk.gov.companieshouse.itemgroupstatusupdater.TestUtils.ERROR_TOPIC;
 import static uk.gov.companieshouse.itemgroupstatusupdater.TestUtils.INVALID_TOPIC;
 import static uk.gov.companieshouse.itemgroupstatusupdater.TestUtils.MAIN_TOPIC;
 import static uk.gov.companieshouse.itemgroupstatusupdater.TestUtils.RETRY_TOPIC;
 
+import java.time.Duration;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.KafkaProducer;
@@ -19,30 +19,21 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.kafka.test.utils.KafkaTestUtils;
 import org.springframework.test.context.ActiveProfiles;
-import java.time.Duration;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 import uk.gov.companieshouse.itemgroupprocessed.ItemGroupProcessed;
-import uk.gov.companieshouse.itemgroupstatusupdater.exception.NonRetryableException;
+import uk.gov.companieshouse.itemgroupstatusupdater.TestUtils;
 
 @SpringBootTest
 @ActiveProfiles("test_main_nonretryable")
-public class ConsumerNonRetryableExceptionTest extends AbstractKafkaIntegrationTest {
-
+class ConsumerInvalidTopicTest extends AbstractKafkaIntegrationTest {
 
     @Autowired
     private KafkaProducer<String, ItemGroupProcessed> testProducer;
     @Autowired
     private KafkaConsumer<String, ItemGroupProcessed> testConsumer;
-
     @Autowired
     private CountDownLatch latch;
-
-    @MockBean
-    private Service service;
 
     @BeforeEach
     public void drainKafkaTopics() {
@@ -50,9 +41,7 @@ public class ConsumerNonRetryableExceptionTest extends AbstractKafkaIntegrationT
     }
 
     @Test
-    void testRepublishToInvalidMessageTopicIfNonRetryableExceptionThrown() throws InterruptedException {
-        //given
-        doThrow(NonRetryableException.class).when(service).processMessage(any());
+    void testPublishToInvalidMessageTopicIfInvalidDataDeserialised() throws InterruptedException {
 
         //when
         testProducer.send(new ProducerRecord<>(MAIN_TOPIC, 0, System.currentTimeMillis(), "key",
@@ -60,6 +49,7 @@ public class ConsumerNonRetryableExceptionTest extends AbstractKafkaIntegrationT
         if (!latch.await(5L, TimeUnit.SECONDS)) {
             fail("Timed out waiting for latch");
         }
+
         ConsumerRecords<?, ?> consumerRecords = KafkaTestUtils.getRecords(testConsumer, 10000L, 2);
 
         //then
@@ -67,6 +57,6 @@ public class ConsumerNonRetryableExceptionTest extends AbstractKafkaIntegrationT
         assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, RETRY_TOPIC)).isZero();
         assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, ERROR_TOPIC)).isZero();
         assertThat(TestUtils.noOfRecordsForTopic(consumerRecords, INVALID_TOPIC)).isEqualTo(1);
-        verify(service).processMessage(new ServiceParameters(ITEM_GROUP_PROCESSED));
     }
+
 }
